Moved experiment running functionality to base class, reduced MLE experiment number + other minor fixes
