+++ tests/tomography/util_test.py
@@ -146,6 +146,18 @@ class TestUtils:
         """
         assert len(_get_tomo_measurements(n_qubits)) == 4**n_qubits
 
+    @pytest.mark.parametrize("remove", [True, False])
+    def test_measurements_remove_trivial(self, remove):
+        """
+        Checks that the trivial I,I measurement is included when required and
+        then removed otherwise
+        """
+        all_meas = _get_tomo_measurements(2, remove_trivial=remove)
+        if not remove:
+            assert "I,I" in all_meas
+        else:
+            assert "I,I" not in all_meas
+
     @pytest.mark.parametrize("n_qubits", [1, 3])
     def test_number_of_required_measurements(self, n_qubits):
         """
